// src/flow/types.ts
export interface FlowDataDemention {
  array: number[] | Float32Array;
  min?: number;
  max?: number;
}

export interface Quantiles {
  q25: number;
  q50: number;
  q75: number;
  q90: number;
}

export interface FlowDataDementionWithQuantiles extends FlowDataDemention {
  quantiles: Quantiles;
}

export interface SeedPoint {
  lon: number;
  lat: number;
}

export interface FlowData {
  u: FlowDataDemention;
  v: FlowDataDemention;
  speed?: FlowDataDementionWithQuantiles;
  width: number;
  height: number;
  bounds: {
    west: number;
    south: number;
    east: number;
    north: number;
  };
  mask?: FlowDataDemention;
  seeds?: SeedPoint[];
}